Move reconcile logic to a different package.
