+++ pkg/refactor/reconcile/reconcile_test.go
@@ -1,4 +1,4 @@
-package service
+package reconcile
 
 import (
 	"fmt"
